backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
protobuf==3.20.3
pyasn1==0.5.0
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
twilio==8.1.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
yarl==1.9.2
